gluster: fix NPE in gluster volumes Tab.
  There is an NPE when we compare null > 0. This happens when
unsyncedEntries is null in volume bricks. So adding a null
check before using this for comparison.

+++ frontend/webadmin/modules/webadmin/src/main/java/org/ovirt/engine/ui/webadmin/ApplicationMessages.java
@@ -126,10 +126,10 @@ public interface ApplicationMessages extends CommonApplicationMessages {
     String onlyAvailableInCompatibilityVersions(String versions);
 
     @DefaultMessage("{0} unsynced entries present")
-    String unSyncedEntriesPresent(double unSyncedEntries);
+    String unSyncedEntriesPresent(int unSyncedEntries);
 
     @DefaultMessage("{0}, {1} unsynced entries present")
-    String brickStatusWithUnSyncedEntriesPresent(String brickStatus, double unSyncedEntries);
+    String brickStatusWithUnSyncedEntriesPresent(String brickStatus, int unSyncedEntries);
 
     @DefaultMessage("{0}, Unsynced entries present - Needs healing")
     String needsGlusterHealingWithVolumeStatus(String volumeStatus);
